refactor: reorganize envSetup and update import paths
