+++ src/configs/envSetup.js
@@ -1,14 +1,10 @@
-import { fileURLToPath } from 'url'
 import dotenv from 'dotenv'
 import path from 'path'
-
-const __filename = fileURLToPath(import.meta.url)
-const __dirname = path.dirname(__filename)
 
 export const loadEnvConfig = () => {
   const NODE_ENV = process.env?.NODE_ENV?.trim() || 'development'
   const envFile = NODE_ENV === 'production' ? '.env.local' : '.env.test.local'
-  const envPath = path.join(__dirname, '..', envFile)
+  const envPath = path.join(envFile)
   const result = dotenv.config({ path: envPath })
 
   if (result.error) throw new Error(`Failed to load ${envFile}: ${result.error.message}`)
